+++ src/components/EditableLabel.vue
@@ -2,9 +2,13 @@
 import { useTextareaAutosize, useVModel } from "@vueuse/core";
 import { ref } from "vue";
 
-const props = defineProps<{
-  modelValue: string;
-}>();
+const props = withDefaults(
+  defineProps<{
+    modelValue: string;
+    textClass?: string;
+  }>(),
+  { textClass: "text-base font-semibold leading-6 text-gray-900" },
+);
 
 const emit = defineEmits(["update:modelValue", "update-value"]);
 
@@ -37,6 +41,7 @@ function onBlur() {
     @focus="onFocus()"
     @blur="onBlur()"
     :spellcheck="spellcheck"
-    class="-mx-3 w-full resize-none rounded-md border-0 text-base font-semibold leading-6 text-gray-900 ring-0 ring-inset hover:ring-1 focus:ring-2 focus:ring-inset"
+    class="-mx-3 w-full resize-none rounded-md border-0 ring-0 ring-inset hover:ring-1 focus:ring-2 focus:ring-inset"
+    :class="textClass"
   />
 </template>